% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expQ2-package.R
\docType{package}
\name{expQ2-package}
\alias{expQ2-package}
\alias{expQ2}
\title{Non-negative vector * exp of rate matrix}
\description{
v exp(Q) by uniformisation or scaling and squaring, or exp(Q) by scaling and squaring.

\tabular{ll}{ 
Package: \tab expQ2\cr 
Type: \tab Package\cr 
Version: \tab 1.1\cr 
Date: \tab March 03, 2022\cr 
License: \tab CC0 \cr 
LazyLoad: \tab yes\cr 
}
}
\note{
This software package is based on the R package {expQ} developed by 
Chris Sherlock (https://github.com/ChrisGSherlock/expQ). This alternate 
version is maintained by scientists at the NOAA Fisheries Pacific Islands 
Fisheries Science Center and should be considered a fundamental research 
communication. The recommendations and conclusions presented here are those 
of the authors and this software should not be construed as official 
communication by NMFS, NOAA, or the U.S. Dept. of Commerce. In addition, 
reference to trade names does not imply endorsement by the National Marine 
Fisheries Service, NOAA. While the best efforts have been made to insure the 
highest quality, tools such as this are under constant development and are 
subject to change.
}
\references{
Sherlock, C. (2021). Direct statistical inference for finite 
Markov jump processes via the matrix exponential. Computational Statistics, 
36(4), 2863-2887.
}
\author{
Chris Sherlock and Devin Johnson

Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}
